+++ packages/vendure-plugin-google-cloud-tasks/src/cloud-tasks.plugin.ts
@@ -2,10 +2,11 @@ import {
   PluginCommonModule,
   RuntimeVendureConfig,
   VendurePlugin,
+  Logger,
 } from '@vendure/core';
 import { CloudTasksJobQueueStrategy } from './cloud-tasks-job-queue.strategy';
 import { CloudTasksHandler } from './cloud-tasks.handler';
-import { CloudTaskOptions, ROUTE } from './types';
+import { CloudTaskOptions } from './types';
 import { JobRecord } from '@vendure/core/dist/plugin/default-job-queue-plugin/job-record.entity';
 
 @VendurePlugin({
@@ -26,6 +27,13 @@ export class CloudTasksPlugin {
 
   static init(options: CloudTaskOptions): typeof CloudTasksPlugin {
     this.options = options;
+    if (this.options?.createTaskRetries > 20) {
+      this.options.createTaskRetries = 20;
+      Logger.warn(
+        `createTaskRetries can be set to a maximum of 20 retries. This is to avoid too many stacked create task retries`,
+        this.loggerCtx
+      );
+    }
     return CloudTasksPlugin;
   }
 }